refactoring method to metrics/utils.py
