+++ src/ragas/metrics/__init__.py
@@ -1,5 +1,3 @@
-from datasets import Dataset
-
 from ragas.metrics._answer_correctness import AnswerCorrectness, answer_correctness
 from ragas.metrics._answer_relevance import AnswerRelevancy, answer_relevancy
 from ragas.metrics._answer_similarity import AnswerSimilarity, answer_similarity
@@ -17,36 +15,6 @@ from ragas.metrics._context_recall import ContextRecall, context_recall
 from ragas.metrics._context_relevancy import ContextRelevancy, context_relevancy
 from ragas.metrics._faithfulness import Faithfulness, faithfulness
 from ragas.metrics.critique import AspectCritique
-from ragas.metrics.base import Metric
-from ragas.validation import validate_evaluation_modes
-
-ALL_METRICS = [
-    answer_correctness,
-    faithfulness,
-    answer_similarity,
-    context_precision,
-    context_utilization,
-    context_recall,
-    context_relevancy,
-    answer_relevancy,
-    context_entity_recall,
-]
-
-
-def get_available_metrics(ds: Dataset) -> list[Metric]:
-    """
-    Get the available metrics for the given dataset.
-    E.g. if the dataset contains ("question", "answer", "contexts") columns, 
-    the available metrics are those that can be evaluated in [qa, qac, qc] mode.
-    """
-    available_metrics = []
-    for metric in ALL_METRICS:
-        try:
-            validate_evaluation_modes(ds, [metric])
-            available_metrics.append(metric)
-        except ValueError:
-            continue
-    return available_metrics
 
 
 __all__ = [
@@ -69,5 +37,4 @@ __all__ = [
     "answer_relevancy",
     "ContextEntityRecall",
     "context_entity_recall",
-    "get_available_metrics",
 ]
